+++ CHANGELOG.md
@@ -3,8 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.1.0]
 ### Added
 - Base functionality
 - Initialize yosys console
 - History of commands
-- Demo
+- Live Demo
+- Pass options to yosys
+- Add multiple files
+- Save files to local storage
+- yosysAfterInit event
+- Display dot files